+++ pkg/querier/queryrange/roundtrip.go
@@ -36,6 +36,7 @@ import (
 
 	"github.com/cortexproject/cortex/pkg/chunk"
 	"github.com/cortexproject/cortex/pkg/chunk/cache"
+	"github.com/cortexproject/cortex/pkg/querier"
 	"github.com/cortexproject/cortex/pkg/tenant"
 	"github.com/cortexproject/cortex/pkg/util"
 	"github.com/cortexproject/cortex/pkg/util/flagext"
@@ -76,12 +77,12 @@ func (cfg *Config) RegisterFlags(f *flag.FlagSet) {
 }
 
 // Validate validates the config.
-func (cfg *Config) Validate() error {
+func (cfg *Config) Validate(qCfg querier.Config) error {
 	if cfg.CacheResults {
 		if cfg.SplitQueriesByInterval <= 0 {
 			return errors.New("querier.cache-results may only be enabled in conjunction with querier.split-queries-by-interval. Please set the latter")
 		}
-		if err := cfg.ResultsCacheConfig.Validate(); err != nil {
+		if err := cfg.ResultsCacheConfig.Validate(qCfg); err != nil {
 			return errors.Wrap(err, "invalid ResultsCache config")
 		}
 	}